requirements.txt
# AutoClaw Tech Stack (v1.0)
playwright>=1.41.0
playwright-stealth>=1.0.6
beautifulsoup4>=4.12.3
pydantic>=2.5.3
instructor>=0.5.2
google-generativeai>=0.4.0
streamlit>=1.31.0
python-telegram-bot>=20.8
SQLAlchemy>=2.0.25
markdownify>=0.11.6
requests>=2.31.0
python-dotenv>=1.0.0
